requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
Django==1.11
django-braces==1.11.0
django-oauth-toolkit==0.12.0
django-rest-framework-social-oauth2==1.0.5
djangorestframework==3.6.2
-e git+https://github.com/mobolic/facebook-sdk.git@cbbe56f00ae7cb838f5dd5e1ca46671a6ec7b0a2#egg=facebook_sdk
oauthlib==2.0.1
packaging==16.8
PyJWT==1.5.0
pyparsing==2.2.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
